// apps/web/lib/api/schema.ts
export interface components {
    schemas: {
// ...
        AuthSuccessResponse: {
            /** Access Token */
            access_token: string;
            /**
             * Token Type
             * @default bearer
             */
            token_type: string;
            user: components["schemas"]["UserResponse"];
            org: components["schemas"]["OrgResponse"];
            role: components["schemas"]["OrgRole"];
        };
        /** DemoLoginResponse */
        DemoLoginResponse: {
            /** Access Token */
            access_token: string;
            /**
             * Token Type
             * @default bearer
             */
            token_type: string;
            user: components["schemas"]["UserResponse"];
            org: components["schemas"]["OrgResponse"];
            role: components["schemas"]["OrgRole"];
            /** Free Runs Remaining */
            free_runs_remaining: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}